Rename `ClaimActionButtons` to `ClaimActions`, to include text input
